requirements.txt
fastapi==0.110.0
mangum==0.17.0
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
nltk==3.8.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
joblib==1.3.2
